# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RentKarRideKar API Gateway",
    "description": "Complete interactive API documentation for RentKarRideKar services including Auth, Vehicle, KYC, Order, Earnings, Referral, Wallet, Disputes, Ratings, Communication, and Admin services.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/admin/offers": {
      "post": {
        "summary": "Create Promo Offer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "value": {
                    "type": "number"
                  },
                  "conditions": {
                    "type": "object"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success"
          }
        },
        "tags": [
          "Admin Service"
        ]
      },
      "get": {
        "summary": "List All Promo Offers",
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Admin Service"
        ]
      }
    },
    "/admin/offers/{id}/toggle": {
      "patch": {
        "summary": "Toggle Promo Offer status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Admin Service"
        ]
      }
    },
    "/offers/validate-code": {
      "post": {
        "summary": "Validate Promo Code",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string"
                  },
                  "originalPrice": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Other Services"
        ]
      }
    },
    "/offers/check-eligibility": {
      "post": {
        "summary": "Check Eligible Offers",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "originalPrice": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Other Services"
        ]
      }
    },
...
  }
}